Move drivetrain related classes to the drivetrain package
