+++ src/org/robockets/stronghold/robot/drivetrain/Drivetrain.java
@@ -1,4 +1,4 @@
-package org.robockets.stronghold.robot.subsystems;
+package org.robockets.stronghold.robot.drivetrain;
 
 import org.robockets.stronghold.robot.RobotMap;
